Adding fixture for enum_downstream_dut_hostname (#20913)
Adding fixture that selects a downstream duthost.
Repo had already fixture for selecting upstream duthost from the available frontend hosts.
This is to cover the gap and add the similar fixture for the downstream type.
This is used by some suites, especially in t2 (e.g. gcu add cluster case #14887).

Signed-off-by: selldinesh <dinesh.sellappan@keysight.com>